requirements.txt
asgiref==3.5.2
certifi==2022.9.14
charset-normalizer==2.1.1
cloudinary==1.29.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==1.0.0
Django==4.1.1
django-cors-headers==3.13.0
django-filter==22.1
django-heroku==0.3.1
djangorestframework==3.13.1
drf-yasg==1.21.3
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
packaging==21.3
psycopg2==2.9.3
pyparsing==3.0.9
pytz==2022.2.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
